Modify some of the CLI for the DSA plugin.
If used with the latest version of the DSA, this will pull model names
from the MONAI Label server.

This passed the girder API URL and a girder token to the clis; these
should be used in the label server to set up the dsa assetstore if they
differ from what was last used.

When running the DSA, if an appropriate environment variable is added to
the girder docker, it will locate the MONAI Label server from that
value.  For instance:

```services: girder: environment: SLICER_CLI_WEB_MONAI_LABEL_SERVER:
https://monai.label.server.com:8020```