Improve the efficiency of working with the DSA.
Before, this often queried all of the items, files, or annotations in
the system.  Now it only queries what is required to perform the desired
task.

I've disabled a status check (by having it always return 0 values).  If
this is actually needed, we should add an endpoint to the DSA to do this
efficiently.  The listing endpoints do return counts in their headers.
The slow query is getting the count of images with annotations.

Signed-off-by: David Manthey <david.manthey@kitware.com>