Slicer plugin: Keep showing segmentation in 3D when using sam_3d
In 3D Slicer, when sam_3d result was imported it always removed the closed surface representation, therefore if a user wanted to see the segmentation in 3D then after each sam_3d update the user had to go to "Segment Editor" section and click "Show 3D".

This commit makes the MONAILabel module remember the "Show 3D" state when performing a sam_3d update.